// frontend/src/openapi/models/OptimizerExperimentStatusEnum.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Overbae API
 * Backend API for the Overclaw / Overmind platform.
 *
 * The version of the OpenAPI document: 2.0.0
 *
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * * `scheduled` - Scheduled
 * * `baseline` - Baseline
 * * `evaluating_baseline_outputs` - Evaluating Baseline Outputs
 * * `evaluated_baseline_outputs` - Evaluated Baseline Outputs
 * * `iterating` - Iterating
 * * `evaluating_candidate_outputs` - Evaluating Candidate Outputs
 * * `evaluated_candidate_outputs` - Evaluated Candidate Outputs
 * * `completed` - Completed
 * * `failed` - Failed
 * * `cancelled` - Cancelled
 * * `paused` - Paused
 * @export
 */
export const OptimizerExperimentStatusEnum = {
    scheduled: 'scheduled',
    baseline: 'baseline',
    evaluating_baseline_outputs: 'evaluating_baseline_outputs',
    evaluated_baseline_outputs: 'evaluated_baseline_outputs',
    iterating: 'iterating',
    evaluating_candidate_outputs: 'evaluating_candidate_outputs',
    evaluated_candidate_outputs: 'evaluated_candidate_outputs',
    completed: 'completed',
    failed: 'failed',
    cancelled: 'cancelled',
    paused: 'paused'
} as const;
export type OptimizerExperimentStatusEnum = typeof OptimizerExperimentStatusEnum[keyof typeof OptimizerExperimentStatusEnum];


export function instanceOfOptimizerExperimentStatusEnum(value: any): boolean {
    for (const key in OptimizerExperimentStatusEnum) {
        if (Object.prototype.hasOwnProperty.call(OptimizerExperimentStatusEnum, key)) {
            if (OptimizerExperimentStatusEnum[key as keyof typeof OptimizerExperimentStatusEnum] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OptimizerExperimentStatusEnumFromJSON(json: any): OptimizerExperimentStatusEnum {
    return OptimizerExperimentStatusEnumFromJSONTyped(json, false);
}

export function OptimizerExperimentStatusEnumFromJSONTyped(json: any, ignoreDiscriminator: boolean): OptimizerExperimentStatusEnum {
    return json as OptimizerExperimentStatusEnum;
}

export function OptimizerExperimentStatusEnumToJSON(value?: OptimizerExperimentStatusEnum | null): any {
    return value as any;
}

export function OptimizerExperimentStatusEnumToJSONTyped(value: any, ignoreDiscriminator: boolean): OptimizerExperimentStatusEnum {
    return value as OptimizerExperimentStatusEnum;
}